+++ src/command_modules/azure-cli-appservice/azure/cli/command_modules/appservice/_help.py
@@ -418,12 +418,32 @@ examples:
 
 helps['functionapp devops-build'] = """
 type: group
-short-summary: Azure Function specific integration with Azure DevOps.
+short-summary: (DEPRECATED) Azure Function specific integration with Azure DevOps.
 """
 
 helps['functionapp devops-build create'] = """
 type: command
+short-summary: (DEPRECATED) Create an Azure DevOps pipeline for a function app. Please use the devops-pipeline create command.
+"""
+
+helps['functionapp devops-pipeline'] = """
+type: group
+short-summary: Azure Function specific integration with Azure DevOps.
+"""
+
+helps['functionapp devops-pipeline create'] = """
+type: command
 short-summary: Create an Azure DevOps pipeline for a function app.
+examples:
+    - name: create an Azure Pipeline to a function app.
+      text: >
+          az functionapp devops-pipeline create --functionapp-name FunctionApp
+    - name: create an Azure Pipeline from a Github function app repository.
+      text: >
+          az functionapp devops-pipeline create --github-repository GithubOrganization/GithubRepository --github-pat GithubPersonalAccessToken
+    - name: create an Azure Pipeline with specific Azure DevOps organization and project
+      text: >
+          az functionapp devops-pipeline create --organization-name AzureDevOpsOrganization --project-name AzureDevOpsProject
 """
 
 helps['functionapp identity'] = """